feat: a little interface polish
