+++ src/app/[locale]/borrower/create-market/components/Forms/MarketPolicyForm/index.tsx
@@ -7,7 +7,6 @@ import {
   Switch,
   TextField,
   Typography,
-  useTheme,
 } from "@mui/material"
 import { DesktopDatePicker } from "@mui/x-date-pickers"
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs"
@@ -24,7 +23,6 @@ import ArrowLeftIcon from "@/assets/icons/sharpArrow_icon.svg"
 import { ExtendedSelect } from "@/components/@extended/ExtendedSelect"
 import { HorizontalInputLabel } from "@/components/HorisontalInputLabel"
 import { InputLabel } from "@/components/InputLabel"
-import { NumberTextField } from "@/components/NumberTextfield"
 import { getMaxFixedTermDays } from "@/config/market-duration"
 import {
   mockedAccessControlOptions,
@@ -72,7 +70,6 @@ export const MarketPolicyForm = ({
   isTestnet,
 }: MarketPolicyFormProps) => {
   const { t } = useTranslation()
-  const theme = useTheme()
   const dispatch = useAppDispatch()
   const router = useRouter()
 
@@ -82,7 +79,6 @@ export const MarketPolicyForm = ({
   const maxDate = today.add(maxDays, "days")
 
   const {
-    getValues,
     setValue,
     register,
     formState: { errors },
@@ -94,17 +90,12 @@ export const MarketPolicyForm = ({
   const policyNameWatch = watch("policyName")
   const implementationTypeWatch = watch("implementationType")
   const marketTypeWatch = watch("marketType")
-  const commitmentFeePercentWatch = watch("commitmentFeePercent")
   const accessControlWatch = watch("accessControl")
   const fixedTermEndTimeWatch = watch("fixedTermEndTime")
   const allowClosureBeforeTermWatch = watch("allowClosureBeforeTerm")
   const allowTermReductionWatch = watch("allowTermReduction")
 
   const isFixedTerm = marketTypeWatch === "fixedTerm"
-  const isRevolving = implementationTypeWatch === "revolving"
-  const hasCommitmentFeeValue =
-    commitmentFeePercentWatch !== undefined &&
-    !Number.isNaN(commitmentFeePercentWatch)
 
   const disableFields = !(
     policyWatch === "createNewPolicy" || policyWatch === ""
@@ -121,7 +112,6 @@ export const MarketPolicyForm = ({
     marketTypeWatch.trim() !== "" &&
     !!accessControlWatch &&
     accessControlWatch.trim() !== "" &&
-    (!isRevolving || (hasCommitmentFeeValue && !errors.commitmentFeePercent)) &&
     !errors.fixedTermEndTime
 
   const isFixedFormValid =
@@ -135,7 +125,6 @@ export const MarketPolicyForm = ({
     marketTypeWatch.trim() !== "" &&
     !!accessControlWatch &&
     accessControlWatch.trim() !== "" &&
-    (!isRevolving || (hasCommitmentFeeValue && !errors.commitmentFeePercent)) &&
     !!fixedTermEndTimeWatch &&
     !errors.fixedTermEndTime
 
@@ -259,34 +248,6 @@ export const MarketPolicyForm = ({
             disabled={disableFields}
           />
         </InputLabel>
-
-        {isRevolving && (
-          <InputLabel
-            label={t("createNewMarket.policy.commitmentFee.label")}
-            subtitle={t("createNewMarket.policy.commitmentFee.explainer")}
-          >
-            <NumberTextField
-              min={0}
-              max={100}
-              decimalScale={2}
-              label={t("createNewMarket.policy.commitmentFee.placeholder")}
-              value={getValues("commitmentFeePercent")}
-              error={Boolean(errors.commitmentFeePercent)}
-              helperText={errors.commitmentFeePercent?.message}
-              endAdornment={
-                <Typography variant="text2" sx={{ color: COLORS.santasGrey }}>
-                  {t("createNewMarket.policy.commitmentFee.chip")}
-                </Typography>
-              }
-              onValueChange={(v) => {
-                setValue("commitmentFeePercent", v.floatValue as number, {
-                  shouldTouch: true,
-                  shouldValidate: true,
-                })
-              }}
-            />
-          </InputLabel>
-        )}
       </Box>
 
       {isFixedTerm && (
